[Feral] Fix convoke in single target without dire fixation
Missing a talent check for dire fixation. Additionally removed some unnecessary lines and added pvp badge to apl. Put in a fix where shadowmeld or prowl would be casted expecting a rake follow-up in situations where the snapshot would be equal (rake was in pandemic) but rake would not be casted, as the snapshot wasn't being increased.